+++ src/common/reposetoryNET.h
@@ -4,7 +4,7 @@
 #include <sys/file.h>
 #include "define.h"
 
-int rGetNextNET(char *HostName, unsigned int LotNr,struct ReposetoryHeaderFormat *ReposetoryHeader, char htmlbuffer[], char imagebuffebuffer[], unsigned long int *radress, unsigned int FilterTime, unsigned int FileOffset,char subname[]);
+int rGetNextNET(char *HostName, unsigned int LotNr,struct ReposetoryHeaderFormat *ReposetoryHeader, char htmlbuffer[], char imagebuffebuffer[], unsigned int *radress, unsigned int FilterTime, unsigned int FileOffset,char subname[]);
 
 int readHTMLNET_toHost(char *subname, unsigned int DocID, char *text, unsigned int len,struct ReposetoryHeaderFormat *ReposetoryHeader, char HostName[]);
 
@@ -15,7 +15,7 @@ int DIReadNET2 ( struct DocumentIndexFormat *DocumentIndexPost, int DocID,char s
 
 void closeNET ();
 
-unsigned long int DIGetIp (char *HostName, unsigned int DocID,char subname[]);
+unsigned int DIGetIp (char *HostName, unsigned int DocID,char subname[]);
 
 
 void setLastIndexTimeForLotNET(char *HostName, int LotNr,char subname[]);
